Fixed: The TypeTraits definition for a bool was inconsistant in how it was reading and writing bool values to/from strings and streams.  In 1.8 OpenSG read/wrote bools to strings as the values "TRUE" or "FALSE".
The TypeTraits getFromCString is reading values from string for bools as either "TRUE" or "FALSE" case insensitive, but the putToString function was not overriden from the TypeTraitsBase definition, so bools were put into the string as "0" or "1".  This was fixed by overriding the putToString and putToStream functions for bool TypeTraits to write "TRUE" or "FALSE".